+++ lib/absinthe/federation/schema.ex
@@ -59,7 +59,8 @@ defmodule Absinthe.Federation.Schema do
   def pipeline(pipeline) do
     Pipeline.insert_after(pipeline, TypeImports, [
       __MODULE__.Phase.AddFederatedTypes,
-      __MODULE__.Phase.AddFederatedDirectives
+      __MODULE__.Phase.AddFederatedDirectives,
+      __MODULE__.Phase.HideResolveReferenceFields
       # __MODULE__.Phase.Validation.KeyFieldsMustExist,
       # __MODULE__.Phase.Validation.KeyFieldsMustBeValidWhenExtends
     ])